% Generated by roxygen2 (4.0.2): do not edit by hand
\name{authenticateUser}
\alias{authenticateUser}
\title{Authenticate at server.}
\usage{
authenticateUser(email = NULL, password = NULL, verbosity = NULL)
}
\arguments{
\item{email}{[\code{character(1)}]\cr
Your e-mail address at OpenMl server. Default is set via \code{\link{setOMLConfig}}.}

\item{password}{[\code{character(1)}]\cr
Your password at OpenML server. Default is set via \code{\link{setOMLConfig}}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{character(1)}]. Session hash for further communication.
}
\description{
Required if you want to upload anything.
}

