% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clearOMLCache}
\alias{clearOMLCache}
\title{Clear cache directories}
\usage{
clearOMLCache()
}
\description{
Delete all cached objects and recreate cache directories.
}
\examples{
\dontrun{
  clearOMLCache()
}
}
