% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertOMLDataSetToMlr.R
\name{convertOMLDataSetToMlr}
\alias{convertOMLDataSetToMlr}
\title{Convert an  data set to mlr.}
\usage{
convertOMLDataSetToMlr(obj, task.type = NULL,
  target = obj$desc$default.target.attribute,
  ignore.flagged.attributes = TRUE, drop.levels = TRUE, verbosity = NULL)
}
\arguments{
\item{obj}{[\code{\link{OMLDataSet}}]\cr
The object that should be converted.}

\item{task.type}{[\code{character(1)}]\cr
As we only pass the data set, we need define the task type manually.
Possible are : \dQuote{Supervised Classification}, \dQuote{Supervised Regression},
\dQuote{Survival Analysis}
Default is to guess it from the target column in the data set, if that is a factor,
we choose classification, if it is numberic we choose regression.}

\item{target}{[\code{character}]\cr
The target for the classification/regression task.
Default is the \code{default.target.attribute} of the \code{DataSetDescription}.}

\item{ignore.flagged.attributes}{[\code{logical(1)}]\cr
Should those features that are listed in the data set description slot \dQuote{ignore.attribute}
be removed?
Default is \code{TRUE}.}

\item{drop.levels}{[\code{logical(1)}]\cr
Drop empty factor levels in data?
Default is \code{TRUE}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{\link[mlr]{Task}}].
}
\description{
Convert an  data set to mlr.
}

