% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOMLMlrRunToBMR.R
\name{convertOMLMlrRunToBMR}
\alias{convertOMLMlrRunToBMR}
\title{Convert \code{OMLMlrRun}s to a \code{BenchmarkResult}.}
\usage{
convertOMLMlrRunToBMR(...)
}
\arguments{
\item{...}{[\code{\link{OMLMlrRun}}]\cr
One or more \code{\link{OMLMlrRun}s}}
}
\value{
[\code{\link[mlr]{BenchmarkResult}}].
}
\description{
Converts one or more \code{\link{OMLMlrRun}}s to a \code{\link[mlr]{BenchmarkResult}}.
}
\seealso{
Other run-related functions: \code{\link{convertOMLRunToBMR}},
  \code{\link{deleteOMLObject}}, \code{\link{getOMLRun}},
  \code{\link{listOMLRuns}},
  \code{\link{makeOMLRunParameter}},
  \code{\link{makeOMLRun}}, \code{\link{tagOMLObject}},
  \code{\link{uploadOMLRun}}
}
\concept{run-related functions}
