% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOMLSetupParamsToDT.R
\name{convertOMLRunEvalsToDT}
\alias{convertOMLRunEvalsToDT}
\title{Extract Parameters from an OpenML run into a flat structure}
\usage{
convertOMLRunEvalsToDT(run.evals, drop.constant = TRUE)
}
\arguments{
\item{run.evals}{[\code{data.frame}]\cr
Result of caling listOMLRunEvaluations(..., setup = TRUE).}
}
\value{
[\code{\link{data.table}}].
}
\description{
Extract Parameters from an OpenML run into a flat structure
}
\seealso{
Other run-related functions: \code{\link{convertOMLMlrRunToBMR}},
  \code{\link{convertOMLRunToBMR}},
  \code{\link{deleteOMLObject}}, \code{\link{getOMLRun}},
  \code{\link{listOMLRuns}},
  \code{\link{makeOMLRunParameter}},
  \code{\link{makeOMLRun}}, \code{\link{tagOMLObject}},
  \code{\link{uploadOMLRun}}
}
\concept{run-related functions}
