% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deleteOMLFlow.R
\name{deleteOMLFlow}
\alias{deleteOMLFlow}
\title{Delete an OpenML flow.}
\usage{
deleteOMLFlow(flow.id, verbosity = NULL)
}
\arguments{
\item{flow.id}{[\code{integer(1)}]\cr
The flow ID. Of course, you can only delete your own flows.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\description{
This will delete one of your uploaded \code{\link{OMLFlow}s}.
}

