% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deleteOMLRun.R
\name{deleteOMLRun}
\alias{deleteOMLRun}
\title{Delete an OpenML run.}
\usage{
deleteOMLRun(run.id, session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{run.id}{[\code{integer(1)}]\cr
The run ID. Of course, you can only delete your own runs.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\description{
This will delete one of your uploaded \code{\link{OMLRun}s}.
}

