% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deleteOpenMLRun}
\alias{deleteOpenMLRun}
\title{Delete an OpenML run.}
\usage{
deleteOpenMLRun(run.id, session.hash,
  show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{run.id}{[\code{integer(1)}]\cr
The run ID. Of course, you can only delete your own runs.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\description{
This will delete one of your uploaded \code{\link{OpenMLRun}s}.
}

