% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadOMLTask}
\alias{downloadOMLTask}
\title{Download a task from the OpenML repository.}
\usage{
downloadOMLTask(id, session.hash, ignore.cache = FALSE, verbosity = NULL)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
ID number of task on OpenML server, used to retrieve the task.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{ignore.cache}{[\code{logical(1)}]\cr
Ignore a possibly cached version and force
re-download of files? Default is \code{FALSE}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{\link{OMLTask}}]
}
\description{
This function downloads an OpenML task and all associated files from the OpenML repository,
intermediately stores the files on disk and creates an S3 object which completely specifies the task.

Usually there is no reason to set the \code{fetch.*} arguments to \code{FALSE},
as you want all information completely encapsulated in the task object.
}
\examples{
# Download task and access relevant information to start running experiments
\dontrun{
task = downloadOMLTask(id = 1)
show(task)
print(task$type)
print(task$target.features)
print(head(task$data.set$data))
}
}

