% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadOpenMLData}
\alias{downloadOpenMLData}
\title{Download an OpenML data set including its description.}
\usage{
downloadOpenMLData(name, version = 1, dir = tempdir(), clean.up = TRUE,
  show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{name}{[\code{character(1)}]\cr
The name of the data set.}

\item{version}{[\code{integer(1)}]\cr
Version that should be downloaded. If the desired version is not available, the latest version
of the data set is downloaded. Default is \code{1}.}

\item{dir}{[\code{character(1)}]\cr
The directory where intermediate files are stored.
Default is the session's temporary directory.}

\item{clean.up}{[\code{loigcal(1)}]\cr
Should the downloaded files be removed from disk at the end?
Default is \code{TRUE}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\value{
[\code{\link{OpenMLDataSetDescription}}]
}
\description{
Download an OpenML data set including its description.
}

