% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadOpenMLDataSet}
\alias{downloadOpenMLDataSet}
\title{Download an OpenML data set.}
\usage{
downloadOpenMLDataSet(id, ignore.cache = FALSE,
  show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
The id of the data set.}

\item{ignore.cache}{[\code{logical(1)}]\cr
Ignore a possibly cached version and force
re-download of files? Default is \code{FALSE}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\value{
[\code{\link{OpenMLDataSetDescription}}]
}
\description{
Given the id, this function will download the whole data set, including the ARFF and
the description XML from the OpenML server. Note that this does not include data
splits or other task-related information. Tasks can be downloaded with
\code{\link{downloadOpenMLTask}}.
}
\seealso{
\code{\link{toMlr}}, \code{\link{downloadOpenMLTask}}
}

