% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadOpenMLImplementation}
\alias{downloadOpenMLImplementation}
\title{Download an OpenML implementation from the OpenML server through a server API call.}
\usage{
downloadOpenMLImplementation(id, download.source.binary = TRUE,
  ignore.cache = FALSE, show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
The implementation id.}

\item{download.source.binary}{[\code{logical(1)}]\cr
Should source / binary files of the implementation also be downloaded?
They will also be stored in \code{dir}.
Default is \code{TRUE}.}

\item{ignore.cache}{[\code{logical(1)}]\cr
Ignore a possibly cached version and force
re-download of files? Default is \code{FALSE}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\value{
[\code{\link{OpenMLImplementation}}].
}
\description{
Retrieves an implementation for a given id.
}

