% Generated by roxygen2 (4.0.1): do not edit by hand
\name{downloadOpenMLRunResults}
\alias{downloadOpenMLRunResults}
\title{Download OpenML run results from the server.}
\usage{
downloadOpenMLRunResults(id, dir = getwd(), show.info = TRUE,
  clean.up = TRUE)
}
\arguments{
\item{id}{[\code{numeric}]\cr
The task ID.}

\item{dir}{[\code{character(1)}]\cr
The directory where to save the downloaded run xml. The file is called "get_run_id.xml" where "id" is replaced
by the actual id. Default is the working directory.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}

\item{clean.up}{[\code{logical(1)}]\cr
Should the downloaded run xml file be removed at the end?
Default is \code{TRUE}.}
}
\value{
[\code{\link{OpenMLRunResults}}]
}
\description{
Download OpenML run results from the server.
}

