% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadOpenMLTask}
\alias{downloadOpenMLTask}
\title{Download a task from the OpenML repository.}
\usage{
downloadOpenMLTask(id, fetch.data.set.description = TRUE,
  fetch.data.set = TRUE, fetch.data.splits = TRUE, ignore.cache = FALSE,
  show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
ID number of task on OpenML server, used to retrieve the task.}

\item{fetch.data.set.description}{[\code{logical(1)}]\cr
Should the data set description also be downloaded?
Default is \code{TRUE}.}

\item{fetch.data.set}{[\code{logical(1)}]\cr
Should the data set also be downloaded?
Default is \code{TRUE}.}

\item{fetch.data.splits}{[\code{logical(1)}]\cr
Should the data splits (for resampling) also be downloaded?
Default is \code{TRUE}.}

\item{ignore.cache}{[\code{logical(1)}]\cr
Ignore a possibly cached version and force
re-download of files? Default is \code{FALSE}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\value{
\code{\link{OpenMLTask}}.
}
\description{
This function downloads an OpenML task and all associated files from the OpenML repository,
intermediately stores the files on disk and creates an S3 object which completely specifies the task.

Usually there is no reason to set the \code{fetch.*} arguments to \code{FALSE},
as you want all information completely encapsulated in the task object.
}
\examples{
# Download task and access relevant information to start running experiments
\dontrun{
task = downloadOpenMLTask(id = 1)
show(task)
print(task$type)
print(task$target.features)
print(head(task$data.desc$data.set))
}
}

