% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadOpenMLTaskResults}
\alias{downloadOpenMLTaskResults}
\title{Download OpenML task results from the server.}
\usage{
downloadOpenMLTaskResults(id, dir = getwd(),
  show.info = getOpenMLOption("show.info"), clean.up = TRUE)
}
\arguments{
\item{id}{[\code{numeric}]\cr
The task ID.}

\item{dir}{[\code{character(1)}]\cr
The directory where to save the downloaded run xml. The file is called "results_task_id.xml" where "id" is
replaced by the actual id. Default is the working directory.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}

\item{clean.up}{[\code{logical(1)}]\cr
Should the downloaded xml file be removed at the end?
Default is \code{TRUE}.}
}
\value{
[\code{\link{OpenMLTaskResults}}]
}
\description{
Download OpenML task results from the server.
}

