% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCachedOMLDataSetStatus.R
\name{getCachedOMLDataSetStatus}
\alias{getCachedOMLDataSetStatus}
\title{Check status of cached datasets.}
\usage{
getCachedOMLDataSetStatus(show.warnings = TRUE)
}
\arguments{
\item{show.warnings}{[\code{logical(1)}]\cr
Show warning if there are deactivated datasets in cache?
Default is \code{TRUE}.}
}
\value{
[\code{data.frame}]
}
\description{
The caching mechanism is fine, but sometimes you might want to
work on a dataset, which is already cached and has been deactivated in the
meanwhile. This function can be used to determine the status of all cached
datasets.
}

