% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDataQualityNames}
\alias{getDataQualityNames}
\title{Retrieve the names of all available data qualities.}
\usage{
getDataQualityNames(set = "all")
}
\arguments{
\item{set}{[\code{character(1)}]\cr
Either \code{"basic"}, which means only rudimentary data quality names
are retrieved, or \code{"all"}.
Default is \code{"all"}.}
}
\value{
[\code{character}].
}
\description{
Retrieve the names of all available data qualities.
}

