% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOMLConfig}
\alias{getOMLConfig}
\title{Returns a list of OpenML configuration settings}
\usage{
getOMLConfig()
}
\value{
\code{list} of current configuration variables with classs \dQuote{OMLConfig}.
}
\description{
Returns a list of OpenML configuration settings
}
\seealso{
Other conf: \code{\link{configuration}};
  \code{\link{setOMLConfig}}
}

