% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOMLDataQualityNames}
\alias{getOMLDataQualityNames}
\title{Retrieve the names of all available data qualities.}
\usage{
getOMLDataQualityNames(set = "all")
}
\arguments{
\item{set}{[\code{character(1)}]\cr
Either \code{"basic"}, which means only rudimentary data quality names
are retrieved, or \code{"all"}.
Default is \code{"all"}.}
}
\value{
[\code{character}].
}
\description{
Retrieve the names of all available data qualities.
}

