% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOMLEvaluationMeasures}
\alias{getOMLEvaluationMeasures}
\title{Get all names of OpenML evaluation measures.}
\usage{
getOMLEvaluationMeasures(session.hash, verbosity = NULL)
}
\arguments{
\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOpenMLConfig}}.}
}
\description{
Get all names of OpenML evaluation measures.
}

