% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOMLPredictions}
\alias{getOMLPredictions}
\title{Get predictions of an OpenML run.}
\usage{
getOMLPredictions(run, session.hash = getSessionHash(),
  ignore.cache = FALSE, verbosity = NULL)
}
\arguments{
\item{run}{[\code{\link{OMLRun}}]\cr
The OpenML run.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{ignore.cache}{[\code{logical(1)}]\cr
Ignore a possibly cached version and force
re-download of files? Default is \code{FALSE}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}]
}
\description{
This function will return the predictions of a certain OpenML run.
}
\seealso{
\code{\link{OMLRun}}, \code{\link{getOMLRun}}
}

