% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getOMLRun.R
\name{getOMLRun}
\alias{getOMLRun}
\title{Get an OpenML run.}
\usage{
getOMLRun(run.id, get.predictions = FALSE, session.hash = getSessionHash(),
  verbosity = NULL)
}
\arguments{
\item{run.id}{[\code{integer(1)}]\cr
The run's ID.}

\item{get.predictions}{[\code{logical(1)}]\cr
Should the associated predictions be retrieved, too? Default is \code{FALSE}. Note that this does not affect the
downloading of metrics and other information.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{\link{OMLRun}}]
}
\description{
This function downloads an OpenML run including all server and user computed metrics.
}
\seealso{
To retrieve the corresponding predictions: \code{\link{getOMLPredictions}}
}

