% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLRunParList_Class.R
\name{getOMLRunParList}
\alias{getOMLRunParList}
\title{Extract OMLRunParList from run}
\usage{
getOMLRunParList(run)
}
\arguments{
\item{run}{[\code{OMLRun}]\cr
A \code{\link{OMLRun}}}
}
\value{
[\code{OMLRunParList}].
}
\description{
Extracts the seed information as \code{\link{OMLRunParList}} from a \code{\link{OMLRun}}.
}

