% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLRunParList_Class.R
\name{getOMLSeedParList}
\alias{getOMLSeedParList}
\title{Extract OMLSeedParList from run}
\usage{
getOMLSeedParList(run)
}
\arguments{
\item{run}{[\code{OMLRun}]\cr
A \code{\link{OMLRun}}}
}
\value{
[\code{OMLSeedParList}].
}
\description{
Extracts the seed information as \code{\link{OMLSeedParList}} from a \code{\link{OMLRun}}.
}

