% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOMLTaskList}
\alias{getOMLTaskList}
\title{Get table of available tasks OpenML server.}
\usage{
getOMLTaskList(type = 1L, session.hash, verbosity = NULL)
}
\arguments{
\item{type}{[\code{integer(1)}]\cr
The task type you want to list tasks for.
See code \link{getOMLTaskTypeList}.
Default is \code{1}.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned data.frame contains the \code{task_id}, the data set id \code{did},
the \code{status} and some describing data qualities.
}

