% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOpenMLConfig}
\alias{getOpenMLConfig}
\title{Returns a list of OpenML configuration settings}
\usage{
getOpenMLConfig()
}
\value{
\code{list} of current configuration variables with classs \dQuote{OpenMLConfig}.
}
\description{
Returns a list of OpenML configuration settings
}
\seealso{
Other conf: \code{\link{configuration}};
  \code{\link{setOpenMLConfig}}
}

