% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOpenMLDataQualityList}
\alias{getOpenMLDataQualityList}
\title{Get all names of OpenML data qualities.}
\usage{
getOpenMLDataQualityList(show.info = FALSE)
}
\arguments{
\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console? Default is \code{FALSE}.}
}
\description{
Get all names of OpenML data qualities.
}

