% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOpenMLDatasetNames}
\alias{getOpenMLDatasetNames}
\title{Get all names of OpenML data sets which are registered on the server.}
\usage{
getOpenMLDatasetNames()
}
\value{
[\code{character}].
}
\description{
Get all names of OpenML data sets which are registered on the server.
}

