% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOpenMLEvaluationMeasures}
\alias{getOpenMLEvaluationMeasures}
\title{Get all names of OpenML evaluation measures.}
\usage{
getOpenMLEvaluationMeasures(show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\description{
Get all names of OpenML evaluation measures.
}

