% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOpenMLOptions}
\alias{getOpenMLOptions}
\title{Returns a list of OpenML's options}
\usage{
getOpenMLOptions()
}
\value{
[\code{list}].
}
\description{
Returns a list of OpenML's options
}
\seealso{
Other configure: \code{\link{configureOpenML}}
}

