% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOpenMLTaskList}
\alias{getOpenMLTaskList}
\title{Get table of available tasks OpenML server.}
\usage{
getOpenMLTaskList(type = 1L, verbosity = NULL)
}
\arguments{
\item{type}{[\code{integer(1)}]\cr
The task type you want to list tasks for.
See code \link{getOpenMLTaskTypeList}.
Default is \code{1}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOpenMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned data.frame contains the \code{task_id}, the data set id \code{did},
the \code{status} and some describing data qualities.
}

