% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOpenMLTaskTypeList}
\alias{getOpenMLTaskTypeList}
\title{Get table of available tasks OpenML server.}
\usage{
getOpenMLTaskTypeList(verbosity = NULL)
}
\arguments{
\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOpenMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned data.frame contains the \code{task_id}, the data set id \code{did},
the \code{status} and some describing data qualities.
}

