% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authenticateUser.R
\name{getSessionHash}
\alias{getSessionHash}
\title{Get the current session's hash.}
\usage{
getSessionHash()
}
\value{
[\code{character(1)}].
}
\description{
Get the current session's hash.
}
\seealso{
Other config: \code{\link{authenticateUser}};
  \code{\link{configuration}}; \code{\link{getOMLConfig}};
  \code{\link{loadOMLConf}}; \code{\link{setOMLConfig}}
}

