% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTasksForDataSet}
\alias{getTasksForDataSet}
\title{Get task IDs for a certain data set}
\usage{
getTasksForDataSet(name)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
The name of the data set.}
}
\value{
[\code{integer}]. A vector containing the task IDs.
}
\description{
Get task IDs for a certain data set
}

