% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listOMLDataSetQualities.R
\name{listOMLDataSetQualities}
\alias{listOMLDataSetQualities}
\title{List available OpenML qualities with values for given data set.}
\usage{
listOMLDataSetQualities(did, session.hash = getSessionHash(),
  verbosity = NULL, name = NULL)
}
\arguments{
\item{did}{[\code{integer(1)}]\cr
The data set ID}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}

\item{name}{[\code{character}]\cr
Retruns only the data qualities from \dQuote{name} (see also \code{\link{listOMLDataSetQualitiesList}}).
Default is \code{NULL} and uses all available data qualities.}
}
\value{
[\code{data.frame}].
}
\description{
The returned \code{data.frame} contains data set quality \dQuote{name} and value \dQuote{value}.
}
\seealso{
Other list: \code{\link{listOMLDataSetQualitiesList}};
  \code{\link{listOMLDataSets}};
  \code{\link{listOMLEvaluationMeasures}};
  \code{\link{listOMLFlows}};
  \code{\link{listOMLRunResults}}
}

