% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listOMLDataSetQualitiesList.R
\name{listOMLDataSetQualitiesList}
\alias{listOMLDataSetQualitiesList}
\title{List available OpenML qualities names.}
\usage{
listOMLDataSetQualitiesList(verbosity = NULL)
}
\arguments{
\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned \code{data.frame} contains quality name \dQuote{name}.
}
\seealso{
Other list: \code{\link{listOMLDataSetQualities}};
  \code{\link{listOMLDataSets}};
  \code{\link{listOMLEvaluationMeasures}};
  \code{\link{listOMLFlows}};
  \code{\link{listOMLRunResults}}
}

