% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listOMLDataSets.R
\name{listOMLDataSets}
\alias{listOMLDataSets}
\title{List available OpenML data sets.}
\usage{
listOMLDataSets(session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned \code{data.frame} contains the data set id \dQuote{did},
the \dQuote{status} (\dQuote{active},(\dQuote{inactive},(\dQuote{in_preparation})
and describing data qualities.
}
\seealso{
Other list: \code{\link{listOMLEvaluationMeasures}};
  \code{\link{listOMLFlows}};
  \code{\link{listOMLRunResults}}
}

