% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listOMLEstimationProcedures.R
\name{listOMLEstimationProcedures}
\alias{listOMLEstimationProcedures}
\title{List available Estimation Procedures}
\usage{
listOMLEstimationProcedures(verbosity = NULL)
}
\arguments{
\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned \code{data.frame} contains the \code{est.id} and the corresponding
name of the estimation procedure.
}
\seealso{
Other listing functions: \code{\link{listOMLDataSetQualities}};
  \code{\link{listOMLDataSets}};
  \code{\link{listOMLEvaluationMeasures}};
  \code{\link{listOMLFlows}}; \code{\link{listOMLRuns}};
  \code{\link{listOMLTaskTypes}};
  \code{\link{listOMLTasks}}
}

