% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listOMLEvaluationMeasures.R
\name{listOMLEvaluationMeasures}
\alias{listOMLEvaluationMeasures}
\title{List available OpenML evaluation measures.}
\usage{
listOMLEvaluationMeasures(session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The names of all evaluation measures which are used in at least one run are returned
in a \code{data.frame}.
}
\seealso{
Other list: \code{\link{listOMLDataSets}};
  \code{\link{listOMLFlows}};
  \code{\link{listOMLRunResults}}
}

