% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listOMLRunResults.R
\name{listOMLRunResults}
\alias{listOMLRunResults}
\title{List run results of a task.}
\usage{
listOMLRunResults(task.id, verbosity = NULL)
}
\arguments{
\item{task.id}{[\code{integer(1)}]\cr
The task ID.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
Retrieves all run results for a task specified by \code{task.id} and returns
it in a \code{data.frame}. Each row contains, among others, the run id \dQuote{rid},
the setup id \dQuote{sid} and the task.id \dQuote{tid}.
}
\seealso{
Other list: \code{\link{getOMLDataSetQualities}};
  \code{\link{listOMLDataSetQualities}};
  \code{\link{listOMLDataSets}};
  \code{\link{listOMLEvaluationMeasures}};
  \code{\link{listOMLFlows}}
}

