% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listOMLTasks.R
\name{listOMLTasks}
\alias{listOMLTasks}
\title{List available OpenML tasks}
\usage{
listOMLTasks(verbosity = NULL, status = "active")
}
\arguments{
\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}

\item{status}{[\code{character}]\cr
Subsets the results according to the status, can take one or more values of
\code{{"active", "deactivated", "in_preparation"}}.
Default is \code{"active"}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned \code{data.frame} contains the \code{task_id}, the data set id \code{did},
the \code{status} and some describing data qualities.
}
\seealso{
Other listing functions: \code{\link{listOMLDataSetQualities}};
  \code{\link{listOMLDataSets}};
  \code{\link{listOMLEstimationProcedures}};
  \code{\link{listOMLEvaluationMeasures}};
  \code{\link{listOMLFlows}}; \code{\link{listOMLRuns}};
  \code{\link{listOMLTaskTypes}}

Other task related functions: \code{\link{OMLTask}},
  \code{\link{makeOMLTask}};
  \code{\link{convertOMLTaskToMlr}};
  \code{\link{deleteOMLObject}}; \code{\link{getOMLTask}};
  \code{\link{listOMLTaskTypes}}
}

