% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listOMLTaskTypes}
\alias{listOMLTaskTypes}
\title{List available OpenML task types}
\usage{
listOMLTaskTypes(session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}] with columns \dQuote{id} and \dQuote{name}.
}
\description{
List available OpenML task types
}

