% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{loadOMLConf}
\alias{loadOMLConf}
\title{Loads a config file from disk to mem}
\usage{
loadOMLConf(path = "~/.openml/config")
}
\arguments{
\item{path}{[\code{character(1)}]\cr
full path location of the config file to be loaded}
}
\value{
\code{list} of current configuration variables with class \dQuote{OMLConfig}.
}
\description{
Loads a config file from disk to mem
}
\seealso{
Other config: \code{\link{authenticateUser}};
  \code{\link{configuration}}; \code{\link{getOMLConfig}};
  \code{\link{getSessionHash}}; \code{\link{setOMLConfig}}
}

