% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeOMLBibRef}
\alias{OMLBibRef}
\alias{makeOMLBibRef}
\title{Construct OMLBibRef.}
\usage{
makeOMLBibRef(citation, url)
}
\arguments{
\item{citation}{[\code{character(1)}]\cr
Free form reference for this implementation.}

\item{url}{[\code{character(1)}]\cr
URL to an online version of the paper, e.g. PDF.}
}
\description{
Construct OMLBibRef.
}

