% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeOMLDataSet}
\alias{OMLDataSet}
\alias{makeOMLDataSet}
\title{Construct OMLDataSet.}
\usage{
makeOMLDataSet(desc, data, colnames.old, colnames.new)
}
\arguments{
\item{desc}{[\code{\link{OMLDataSetDescription}}]\cr
The data set's description.}

\item{data}{[\code{data.frame}]\cr
The data set.}

\item{colnames.old}{[\code{character}]\cr
The original column names of the data set. These might not be valid names in R!}

\item{colnames.new}{[\code{character}]\cr
New column names that are actually used in R. These are valid and unique and differ from the
original column names only if those are invalid.}
}
\description{
Construct OMLDataSet.
}

