% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLFlow_Class.R
\name{makeOMLFlowParameter}
\alias{OMLFlowParameter}
\alias{makeOMLFlowParameter}
\title{Construct OMLFlowParameter.}
\usage{
makeOMLFlowParameter(name, data.type = NA_character_,
  default.value = NA_character_, description = NA_character_,
  recommended.range = NA_character_)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
The name of the parameter.}

\item{data.type}{[\code{character(1)}]\cr
The data type of the parameter. Should be either integer, numeric, string, vector, matrix or object.}

\item{default.value}{[\code{character(1)}]\cr
The default value of the parameter.}

\item{description}{[\code{character(1)}]\cr
A description of what this parameter does.}

\item{recommended.range}{[\code{character(1)}]\cr
Minimal/maximal value and/or a recommended range of values.}
}
\description{
Construct OMLFlowParameter.
}
\seealso{
Other flow-related functions: \code{\link{deleteOMLObject}},
  \code{\link{getOMLFlow}}, \code{\link{listOMLFlows}},
  \code{\link{makeOMLFlow}}, \code{\link{tagOMLObject}}
}
\keyword{internal}

