% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OMLFlow_Class.R
\name{makeOMLFlowQuality}
\alias{OMLFlowQuality}
\alias{makeOMLFlowQuality}
\title{Construct OMLFlowQuality.}
\usage{
makeOMLFlowQuality(name, value)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Name of the quality.\cr
Suggested: LearnerType (classification, regression, cost sensitive classification, survival analysis, clustering),
HandlesNumericFeatures, HandlesNominalFeatures, HandlesMissingValues, HandlesInstanceWeights, HandlesMultiClass,
HandlesBinaryClass, HandlesSingleClass, HandlesOrderedFeatures, HandlesCostMatrices, HandlesClassWeights,
CanOutputProbabilities (classification+clustering), CanOutputStandardError (regression).}

\item{value}{[\code{character(1)}]\cr
The quality's value. E.g., 'true' or 'false'.}
}
\description{
Construct OMLFlowQuality.
}

