% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeOMLIOData}
\alias{OMLIOData}
\alias{makeOMLIOData}
\title{Construct OMLIOData.}
\usage{
makeOMLIOData(datasets = NULL, files = NULL, evaluations = NULL)
}
\arguments{
\item{datasets}{[\code{data.frame}]\cr An optional \code{data.frame} with columns "did", "name" and "url".
Default is \code{NULL}, which means there are no data sets.}

\item{files}{[\code{data.frame}]\cr An optional \code{data.frame} with columns "did", "name" and "url".
Default is \code{NULL}, which means there are no files.}

\item{evaluations}{[\code{data.frame}]\cr
An optional list of one or more evaluations. Default is \code{NULL}, which means there are
no data sets.}
}
\description{
Construct OMLIOData.
}

