% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLRunParameter_Class.R
\name{makeOMLRunParameter}
\alias{OMLRunParameter}
\alias{makeOMLRunParameter}
\title{Construct OMLRunParameter.}
\usage{
makeOMLRunParameter(name, value, component = NA_character_)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
The name of the parameter.}

\item{value}{[\code{character(1)}]\cr
The value of the parameter.}

\item{component}{[\code{character(1)}]\cr
The implementation name of a component, if the parameter belongs to this component.
This name must match a component of the implementation.}
}
\description{
Construct a run parameter object for a run.
}
\seealso{
Other run related functions: \code{\link{deleteOMLObject}},
  \code{\link{getOMLRun}}, \code{\link{listOMLRuns}},
  \code{\link{makeOMLRun}}, \code{\link{uploadOMLRun}}
}

