% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLRunParList_Class.R
\name{makeOMLRunParList}
\alias{makeOMLRunParList}
\alias{OMLRunParList}
\title{Construct OMLRunParList.}
\usage{
makeOMLRunParList(mlr.lrn, component = NA_character_)
}
\arguments{
\item{mlr.lrn}{[\code{\link[mlr]{Learner}}]\cr
The mlr learner.}

\item{component}{[\code{character}]\cr
If the learner is a (sub-)component of a flow, this component's name.}
}
\value{
A \code{OMLRunParList} which is a list of \code{\link{OMLRunParameter}s}.
}
\description{
Generate a list of OpenML run parameter settings for a given mlr learner.
}
