% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLSeedParList_Class.R
\name{makeOMLSeedParList}
\alias{OMLSeedParList}
\alias{makeOMLSeedParList}
\title{Construct OMLSeedParList}
\usage{
makeOMLSeedParList(seed, prefix = "openml")
}
\arguments{
\item{seed}{[\code{numeric(1)}]\cr
The seed.}

\item{prefix}{[\code{character}]\cr
prefix for seed parameter names.}
}
\value{
A \code{OMLSeedParList} which is a list of \code{\link{OMLRunParameter}s} 
that provide only information about the seed.
}
\description{
Generate a list of OpenML seed parameter settings for a given seed.
}

