% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLStudy_Class.R
\name{makeOMLStudy}
\alias{makeOMLStudy}
\alias{OMLStudy}
\title{OMLStudy.}
\usage{
makeOMLStudy(
  alias,
  name,
  description,
  data.id = NULL,
  task.id = NULL,
  flow.id = NULL,
  run.id = NULL
)
}
\arguments{
\item{alias}{[\code{character}]\cr
The alias of the study.}

\item{name}{[\code{character}]\cr
The name of the study.}

\item{description}{[\code{character}]\cr
The description of the study.}

\item{data.id}{[\code{integer}]\cr
A vector of IDs of the data sets to be included in the study.}

\item{task.id}{[\code{integer}]\cr
A vector of IDs of the tasks to be included in the study.}

\item{flow.id}{[\code{integer}]\cr
A vector of IDs of the flows to be included in the study.}

\item{run.id}{[\code{integer}]\cr
A vector of IDs of the runs to be included in the study.}
}
\value{
[\code{OMLStudy}].
}
\description{
If you create a study through the website \url{https://www.openml.org/new/study}, you can also specify an alias which can be used to access the study.
To see a full list of all elements, please see the
\href{https://docs.openml.org/}{documentation}.
}
\seealso{
Other uploading functions: 
\code{\link{uploadOMLDataSet}()},
\code{\link{uploadOMLFlow}()},
\code{\link{uploadOMLRun}()},
\code{\link{uploadOMLStudy}()}
}
\concept{uploading functions}
