% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeOMLTask}
\alias{OMLTask}
\alias{makeOMLTask}
\title{Construct OMLTask.}
\usage{
makeOMLTask(id, type, pars = list(), target.features = NA_character_,
  data.desc.id, data.set = NULL, estimation.procedure, preds = list(),
  evaluation.measures)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
The ID of this task. Generated by the API.}

\item{type}{[\code{character(1)}]\cr
The task type of this task. Task types can be browsed and created on the OpenML website.}

\item{pars}{[\code{list}]\cr
A set of parameters specific to the task type.}

\item{target.features}{[\code{character}]\cr
The name(s) of the target feature(s). There may be none, one or multiple target features.}

\item{data.desc.id}{[\code{integer(1)}]\cr
The OpenML ID of the data set associated with the task.}

\item{data.set}{[\code{\link{OMLDataSet}}]\cr
The data set including its description.}

\item{estimation.procedure}{[\code{\link{OMLEstimationProcedure}}]\cr
Information on the task's estimation method and the asoociated data splits. Optional.}

\item{preds}{[\code{list}]\cr
A list that contains information on the format of the predictions for the particular task.}

\item{evaluation.measures}{[\code{character}]\cr
The evaluation measure(s) that should be used for optimization.}
}
\description{
Construct OMLTask.
}
\seealso{
\code{\link{OMLDataSetDescription}}
}

