% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeOpenMLBibRef}
\alias{OpenMLBibRef}
\alias{makeOpenMLBibRef}
\title{Construct OpenMLBibRef.}
\usage{
makeOpenMLBibRef(citation, url)
}
\arguments{
\item{citation}{[\code{character(1)}]\cr
Free form reference for this implementation.}

\item{url}{[\code{character(1)}]\cr
URL to an online version of the paper, e.g. PDF.}
}
\description{
Construct OpenMLBibRef.
}

