% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeOpenMLDataSetDescription}
\alias{OpenMLDataSetDescription}
\alias{makeOpenMLDataSetDescription}
\title{Construct OpenMLDataSetDescription.}
\usage{
makeOpenMLDataSetDescription(id, name, version, description, format,
  creator = NA_character_, contributor = NA_character_,
  collection.date = NA_character_, upload.date, language = NA_character_,
  licence = NA_character_, url, row.id.attribute = NA_character_,
  md5.checksum = NA_character_, data.set,
  original.col.names = NA_character_, new.col.names = NA_character_)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
ID, autogenerated by the server.}

\item{name}{[\code{character(1)}]\cr
The  name of the data set.}

\item{version}{[\code{character(1)}]\cr
Version of the data set, added by the server.}

\item{description}{[\code{character(1)}]\cr
Description of the data set, given by the uploader.}

\item{format}{[\code{character(1)}]\cr
Format of the data set. Typically, this is "arff".}

\item{creator}{[\code{character}]\cr
The person(s), that created this data set. Optional.}

\item{contributor}{[\code{character}]\cr
People, that contibuted to this version of the data set (e.g., by reformatting). Optional.}

\item{collection.date}{[\code{character(1)}]\cr
The date the data was originally collected. Given by the uploader. Optional.}

\item{upload.date}{[\code{\link[base]{POSIXt}}]\cr
The date the data was uploaded. Added by the server.}

\item{language}{[\code{character(1)}]\cr
Language in which the data is represented. Starts with 1 upper case letter, rest lower case,
e.g. 'English'}

\item{licence}{[\code{character(1)}]\cr
Licence of the data. \code{NA} means: Public Domain or "don't know/care".}

\item{url}{[\code{character(1)}]\cr
Valid URL that points to the data file.}

\item{row.id.attribute}{[\code{character(1)}]\cr
The attribute that represents the row-id column, if present in the dataset. Else \code{NA}.}

\item{md5.checksum}{[\code{character(1)}]\cr
MD5 checksum to check if the dataset is downloaded without corruption.}

\item{data.set}{[\code{data.frame}]\cr
The data set.}

\item{original.col.names}{[\code{character}]\cr
The original column names of the data set. These might not be valid names in R!}

\item{new.col.names}{[\code{character}]\cr
New column names that are actually used in R. These are valid and unique and differ from the original
column names only if those are invalid.}
}
\description{
Construct OpenMLDataSetDescription.
}

