% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeOpenMLIOData}
\alias{OpenMLIOData}
\alias{makeOpenMLIOData}
\title{Construct OpenMLIOData.}
\usage{
makeOpenMLIOData(dataset = NULL, evaluation)
}
\arguments{
\item{dataset}{[\code{list}]\cr An optional list of one or more data set descriptions with
elements "did", "name" and "url".
Default is \code{NULL}, which means there are no data sets.}

\item{evaluation}{[\code{data.frame}]\cr
An optional list of one or more evaluations. Default is \code{NULL}, which means there are
no data sets.}
}
\description{
Construct OpenMLIOData.
}

