% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeOpenMLRun}
\alias{OpenMLRun}
\alias{makeOpenMLRun}
\title{Construct OpenMLRun.}
\usage{
makeOpenMLRun(task.id, implementation.id, error.message = NA_character_,
  parameter.settings = list())
}
\arguments{
\item{task.id}{[\code{character(1)}]\cr
The id of the task.}

\item{implementation.id}{[\code{character(1)}]\cr
The id of the used implementation.}

\item{error.message}{[\code{character(1)}]\cr
Potential error message generated during run.
Default is NA, which means no error occurred.}

\item{parameter.settings}{[named \code{list}]\cr
Optional parameter settings for this run.
Default is empty list.}
}
\description{
Construct OpenMLRun.
}

