% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeOpenMLRunParameter}
\alias{OpenMLRunParameter}
\alias{makeOpenMLRunParameter}
\title{Construct OpenMLRunParameter}
\usage{
makeOpenMLRunParameter(name, value, component = NA_character_)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
The name of the parameter.}

\item{value}{[\code{character(1)}]\cr
The value of the parameter.}

\item{component}{[\code{character(1)}]\cr
   The implementation name of a component, if the parameter belongs to this component.
   This name must match a component of the implementation.}
}
\description{
Construct OpenMLRunParameter
}

