% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeOpenMLTask}
\alias{OpenMLTask}
\alias{makeOpenMLTask}
\title{Construct OpenMLTask.}
\usage{
makeOpenMLTask(id, type, pars = list(), target.features = NA_character_,
  data.desc.id, data.desc = NULL, estimation.procedure, preds = list(),
  evaluation.measures)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
The ID of this task. Generated by the API.}

\item{type}{[\code{character(1)}]\cr
The task type of this task. Task types can be browsed and created on the OpenML website.}

\item{pars}{[\code{list}]\cr
A set of parameters specific to the task type.}

\item{target.features}{[\code{character}]\cr
The name(s) of the target feature(s). There may be none, one or multiple target features.}

\item{data.desc.id}{[\code{integer(1)}]\cr
The OpenML ID of the data set associated with the task.}

\item{data.desc}{[\code{\link{OpenMLDataSetDescription}}]\cr
Information on the data set. Optional.}

\item{estimation.procedure}{[\code{\link{OpenMLEstimationProcedure}}]\cr
Information on the task's estimation method and the asoociated data splits. Optional.}

\item{preds}{[\code{list}]\cr
A list that contains information on the format of the predictions for the particular task.}

\item{evaluation.measures}{[\code{character}]\cr
The evaluation measure(s) that should be used for optimization.}
}
\description{
Construct OpenMLTask.
}
\seealso{
\code{\linkS4class{OpenMLDataSetDescription}}
}

