% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeRunParameterList}
\alias{makeRunParameterList}
\title{Generate a list of OpenML run parameter settings for a given mlr learner.}
\usage{
makeRunParameterList(mlr.lrn, component = NULL)
}
\arguments{
\item{mlr.lrn}{[\code{\link[mlr]{Learner}}]\cr
The mlr learner.}

\item{component}{[\code{character}]\cr
If the learner is a (sub-)component of an implementation, this component's name.}
}
\value{
A list of \code{\link{OpenMLRunParameter}s}.
}
\description{
Generate a list of OpenML run parameter settings for a given mlr learner.
}
\examples{
library(mlr)
lrn = makeLearner("classif.rpart", minsplit = 1)
bagging = makeBaggingWrapper(lrn, bag.iters = 500)

lrn.par.settings = makeRunParameterList(lrn)
lrn.par.settings

bagging.par.settings = makeRunParameterList(bagging)
bagging.par.settings
}

