% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLDataSet_Class.R
\name{OMLDataSet}
\alias{OMLDataSet}
\alias{makeOMLDataSet}
\title{OMLDataSet.}
\usage{
makeOMLDataSet(desc, data, colnames.old, colnames.new, target.features)
}
\arguments{
\item{desc}{[\code{OMLDataSetDescription}]\cr
Data set description.}

\item{data}{[\code{data.frame}]\cr
The data set.}

\item{colnames.old}{[\code{character}]\cr
Names of the features that were uploaded to the server.}

\item{colnames.new}{[\code{character}]\cr
Names of the features that are displayed.}

\item{target.features}{[\code{character}]\cr
Name(s) of the target feature(s).}
}
\value{
[\code{OMLDataSet}]
}
\description{
An \code{OMLDataSet} consists of an \code{OMLDataSetDescription}, a
\code{data.frame} containing the data set, the old and new column names and,
finally, the target features.

The \code{\link{OMLDataSetDescription}} provides information on the data set,
like the ID, name, version, etc. To see a full list of all elements, please see the
\href{https://github.com/openml/website/blob/master/openml_OS/views/pages/api_new/v1/xsd/openml.data.upload.xsd}{XSD}.

The slot \code{colnames.old} contains the original names, i.e., the column names that were
uploaded to the server, while \code{colnames.new} contains the names that you will see when
working with the data in R.
Most of the time, old and new column names are identical. Only if the original names are
not valid, the new ones will differ.

The slot \code{target.features} contains the column name(s) from the \code{data.frame}
of the \code{OMLDataSet} that refer to the target feature(s).
}
\examples{
data(CHFLS, package = "HSAUR3")
dsc <- "The Chinese Health and Family Life Survey sampled $60$ villages and urban 
neighborhoods chosen in such a way as to represent the full geographical and 
socioeconomic range of contemporary China.
This data set is taken from R package HSAUR3 (version 1.0-5)"
cit <- "William L. Parish, Edward O. Laumann, Myron S. Cohen, Suiming Pan, Heyi Zheng, 
Irving Hoffman, Tianfu Wang, and Kwai Hang Ng. (2003), Population-Based Study of 
Chlamydial Infection in China: A Hidden Epidemic. Journal of the American Medican 
Association, 289(10), 1265–1273."
desc_chfls <- makeOMLDataSetDescription(name = "CHFLS", 
    description = dsc, 
    creator = "Chinese Health and Family Life Survey",
    contributor = "Torsten Hothorn and Brian S. Everitt",
    collection.date = as.POSIXct("2015-07-28"),
    language = "English",
    licence = "GPL-2", 
    url = "https://cran.r-project.org/web/packages/HSAUR3/",
    default.target.attribute = NA_character_,
    citation = cit, 
    original.data.url = "http://www.spc.uchicago.edu/prc/chfls.php",
    tags = "R")

chfls_oml <- makeOMLDataSet(desc = desc_chfls, 
                            data = CHFLS, 
                            colnames.old = colnames(CHFLS), 
                            colnames.new = colnames(CHFLS),
                            target.features = c("R_income", "R_happy", "A_income"))
}
\seealso{
Other dataset related functions: \code{\link{OMLDataSetDescription}},
  \code{\link{convertMlrTaskToOMLDataSet}},
  \code{\link{convertOMLDataSetToMlr}},
  \code{\link{deleteOMLObject}},
  \code{\link{getOMLDataSet}},
  \code{\link{listOMLDataSets}},
  \code{\link{tagOMLObject}},
  \code{\link{uploadOMLDataSet}}
}

