% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLIOData_Class.R
\name{OMLIOData}
\alias{OMLIOData}
\title{OMLIOData.}
\description{
Objects of class \code{OMLIOData} are used to return information about all inputs 
  and outputs of \code{\link{OMLRun}s}. Each object may contain up to three \code{data.frames},
  containing information on \code{datasets} ("did", "name", "url"), \code{files} ("did", "name", "url")
  and/or \code{evaluations}.
}

