% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{OpenMLBibRef}
\alias{OpenMLBibRef}
\alias{OpenMLBibRef-class}
\title{OpenMLBibRef}
\description{
This class of objects contains information on a bibliographical reference for an implementation.
}
\details{
Objects can be created by calls of the form \code{OpenMLBibRef(...)}.
The objects contain information on ... .
}
\section{Slots}{

 \describe{
   \item{\code{citation}}{[\code{character}]\cr
   Free form reference for this implementation.}
   \item{\code{url}}{[\code{character}]\cr
   URL to an online version of the paper, e.g. PDF.}
 }
}

