% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{OpenMLData}
\alias{OpenMLData}
\alias{OpenMLData-class}
\title{OpenMLData}
\description{
This class of objects contains information on datasets and/or evaluations. This class is used in
\code{\link{OpenMLRun}s}.
}
\details{
Objects can be created by calls of the form \code{OpenMLData(...)}.
The objects contain information on ... .
}
\section{Slots}{

 \describe{
   \item{\code{dataset}}{[\code{list}]\cr
   A list of short data set descriptions with elements "did", "name" and "url". This list is possibly empty.}
   \item{\code{evaluation}}{[\code{list}]\cr
   A list of evaluations. This list is possibly empty.}
 }
}

