% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{OpenMLRunResults}
\alias{OpenMLRunResults}
\alias{OpenMLRunResults-class}
\title{OpenMLRunResults}
\description{
This class of objects contains information on the results of an \code{\link{OpenMLRun}}.
The objects contain information on...
}
\section{Slots}{

  \describe{
    \item{\code{run.id}}{[\code{numeric(1)}]\cr
      ID of the run. Added by server. Ignored when uploading a run.}
    \item{\code{uploader}}{[\code{numeric(1)}]\cr
      ID of the user that uploaded the run. Added by server. Ignored when uploading a run.}
    \item{\code{task.id}}{[\code{numeric(1)}]\cr
      ID of the task that is solved in this run. This ID is given in the task description.}
    \item{\code{implementation.id}}{[\code{character(1)}]\cr
      ID of the implementation used to solve the task. Returned by the API when you first upload the
      implementation, or given in the implementation description when you download an existing implementation. }
    \item{\code{setup.id}}{[\code{numeric(1)}]\cr
      Unique ID of the used setup. Ignored when uploading a run (i.e., it will be searched based
      on the parameter settings).}
    \item{\code{error.message}}{[\code{character(1)}]\cr
      Whenever an error occurs during the run, this can be reported here.}
    \item{\code{parameter.setting}}{[\code{list}]\cr
      A list of \code{\link{OpenMLRunParameter}s} containing information on the parameter settings.}
    \item{\code{input.data}}{[\code{\link{OpenMLData}}]\cr
      All data that served as input for the run. Added by server. Ignored when uploading.}
    \item{\code{output.data}}{[\code{\link{OpenMLData}}]\cr
      All data that was the output of this run, i.e., predictions, evaluation scores.
      Most of this will be added by the server, but users can also provide evaluation scores for their
      own evaluation measures.}
  }
}

