% Generated by roxygen2 (4.0.2): do not edit by hand
\name{printDescription}
\alias{printDescription}
\title{Print the description of an OpenML object.}
\usage{
printDescription(x)
}
\arguments{
\item{x}{The object whose description should be printed. This is typically an \code{\link{OMLDataSet}},
an \code{\link{OMLDataSetDescription}}, an \code{\link{OMLImplementation}} or an
\code{\link{OMLTask}}. In the latter case, the description of the corresponding
data set is printed.}
}
\description{
Print the description of an OpenML object.
}

