% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resetOpenMLRun}
\alias{resetOpenMLRun}
\title{Reset an OpenMLRun.}
\usage{
resetOpenMLRun(run.id, session.hash, show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{run.id}{[\code{integer(1)}]\cr
The run ID. Note, that you can only reset your own runs.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\description{
This will reset the status of one of your uploaded \code{\link{OpenMLRun}s} to
  "unprocessed", meaning that the server's evaluation engine will pick up on it again. This is
  useful if a temporary server problem occured so that your run has not been processed correctly.
}

