% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runMultipleLearnersOnTask.R
\name{runMultipleLearnersOnTask}
\alias{runMultipleLearnersOnTask}
\title{Run multiple mlr learners on OpenML task.

Run task with a specified learners from mlr and produce predictions.}
\usage{
runMultipleLearnersOnTask(task, learner, remove.const.feats = TRUE,
  verbosity = NULL, auto.upload = TRUE, ...)
}
\arguments{
\item{task}{[\code{\link{OMLTask}}]\cr
An OpenML task.}

\item{learner}{[list of \code{\link[mlr]{Learner}}]\cr
Learner from package mlr to run the task.}

\item{remove.const.feats}{[\code{logical(1)}]\cr
Should constant features be removed?
Default is \code{TRUE}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}

\item{auto.upload}{[\code{logical(1)}]\cr
  Checks whether an \code{\link{OMLFlow}} object containing the passed \code{learner}
  was already uploaded to the server.

  If it has not been found on the server and \code{auto.upload = TRUE},
  a new \code{implementation.id} is assigned and the \code{\link{OMLFlow}} is
  automatically uploaded. If the \code{learner} was already uploaded, the
  \code{implementation.id} of the respective \code{\link{OMLFlow}} is used.

  If \code{auto.upload = FALSE}, only the \code{implementation.id} of an
  already uploaded \code{learner} is used and an error is returned if the
  \code{learner} was not found on the server.}

\item{...}{[any]\cr
Further arguments that are passed to \code{\link[mlr]{removeConstantFeatures}}.}
}
\value{
[\code{OMLMlrRun}], an \code{\link{OMLRun}} with an additional slot \code{mlr.resample.result}.
}
\description{
Run multiple mlr learners on OpenML task.

Run task with a specified learners from mlr and produce predictions.
}
\seealso{
\code{\link{getOMLTask}}, \code{\link[mlr]{makeLearner}}
}

