% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runSQLQuery}
\alias{runSQLQuery}
\title{Run an SQL query on the OpenML DB and retrieve results.}
\usage{
runSQLQuery(query, simplify = TRUE, verbosity = NULL)
}
\arguments{
\item{query}{[\code{character(1)}]\cr
The SQL query.}

\item{simplify}{[\code{logical(1)}]\cr
If there is only one column in the resulting table,
directly return the corresponding vector?
Default is \code{TRUE}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}]. The results as a table.
}
\description{
Most users should not need to use this, only a last
measure to access all information on the server.
}

