% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTaskMlr.R
\name{runTaskMlr}
\alias{runTaskMlr}
\title{Run mlr learner on OpenML task.}
\usage{
runTaskMlr(task, learner, verbosity = NULL, seed = 1, ...)
}
\arguments{
\item{task}{[\code{\link{OMLTask}}]\cr
An OpenML task.}

\item{learner}{[\code{\link[mlr]{Learner}}]\cr
Learner from package mlr to run the task.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}

\item{seed}{[\code{numeric(1)}]\cr
Set a seed to reproduce this run. 
Default is \code{1}.}

\item{...}{[any]\cr
Further arguments that are passed to \code{\link{convertOMLTaskToMlr}}.}
}
\value{
[\code{OMLMlrRun}], an \code{\link{OMLRun}} with additional slots: \code{mlr.benchmark.result} and \code{flow}.
}
\description{
Run task with a specified learner from mlr and produce predictions.
}
\seealso{
\code{\link{getOMLTask}}, \code{\link[mlr]{makeLearner}}
}

