% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveOMLConfig.R
\name{saveOMLConfig}
\alias{saveOMLConfig}
\title{Saves a list of OpenML configuration settings to file}
\usage{
saveOMLConfig(conf = list(), ..., path = "~/.openml/config",
  overwrite = FALSE)
}
\arguments{
\item{conf}{[\code{list}]\cr
Named list of configuration values.}

\item{...}{[any]\cr
Another possibility to pass configuration options.}

\item{path}{[\code{character(1)}]\cr
Path to OpenML config file. Default is \dQuote{~/.openml/config}.}

\item{overwrite}{[\code{logical(1)}]\cr
Should an existing file be overwritten? Default is \code{FALSE}.}
}
\description{
Saves a list of OpenML configuration settings to file
}
\seealso{
Other config: \code{\link{configuration}};
  \code{\link{getOMLConfig}}; \code{\link{loadOMLConfig}};
  \code{\link{setOMLConfig}}
}

