% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setOMLConfig.R
\name{setOMLConfig}
\alias{setOMLConfig}
\title{Set and overwrite configuration settings}
\usage{
setOMLConfig(..., conf = NULL)
}
\arguments{
\item{...}{[\code{ANY}]\cr
Named configuration parameters. Overwrites parameters in \code{conf}, if provided.}

\item{conf}{[\code{OMLConfig} or \code{list}]\cr
List of configuration parameters as returned by \code{\link{getOMLConfig}}.}
}
\value{
Invisibly returns a list of configuration settings.
}
\description{
Set and overwrite configuration settings
}
\seealso{
Other config: \code{\link{configuration}};
  \code{\link{getOMLConfig}}; \code{\link{loadOMLConfig}};
  \code{\link{saveOMLConfig}}
}

