% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setOpenMLConfig}
\alias{setOpenMLConfig}
\title{Set and overwrite configuration settings}
\usage{
setOpenMLConfig(conf = list(), ...)
}
\arguments{
\item{conf}{[\code{OpenMLConfig} or \code{list}]\cr
List of configuration parameters as returned by \code{\link{loadConfig}} or \code{\link{getConfig}}.}

\item{...}{[\code{ANY}]\cr
Named configuration parameters. Overwrites parameters in \code{conf}, if provided.}
}
\value{
Invisibly returns a list of configuration settings.
}
\description{
Set and overwrite configuration settings
}
\seealso{
Other conf: \code{\link{configuration}};
  \code{\link{getOpenMLConfig}}
}

