% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toMlr}
\alias{toMlr}
\alias{toMlr.OpenMLDataSetDescription}
\alias{toMlr.OpenMLTask}
\title{Convert an OpenML object to mlr.}
\usage{
toMlr(obj, target, remove.target.NAs)

\method{toMlr}{OpenMLTask}(obj,
  target = obj$data.desc$default.target.attribute, remove.target.NAs = TRUE)

\method{toMlr}{OpenMLDataSetDescription}(obj,
  target = obj$default.target.attribute, remove.target.NAs = TRUE)
}
\arguments{
\item{obj}{[\code{\link{OpenMLDataSetDescription}} | \code{\link{OpenMLTask}}]\cr
The object that should be converted. Required.}

\item{target}{[\code{character}]\cr
The target for the classification/regression task. Default is the
\code{default.target.attribute} of the \code{DataSetDescription}.}

\item{remove.target.NAs}{[\code{logical(1)}]\cr
Should rows with missing target values be removed? Default is \code{TRUE}. Note, that
the function might fail if you set this to \code{FALSE}.}
}
\value{
[\code{\link[mlr]{Task}}]
}
\description{
This function converts an \code{\link{OpenMLDataSetDescription}}
  or an \code{\link{OpenMLTask}} into an mlr \code{\link[mlr]{Task}}. In
  the latter case, a list is returned with the following elements:\cr
  \code{mlr.task} -- the \code{\link[mlr]{Task}},\cr \code{mlr.rin} -- a
  \code{\link[mlr]{ResampleInstance}} which was defined by the server,\cr
  \code{mlr.measures} -- a list of \code{\link[mlr]{Measure}s} to optimize
  for,\cr \code{orig.lvls} -- a \code{character} vector containing all
  original (possibly empty) levels of the target feature. Empty levels are
  removed from the data set during conversion.
}

