% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uploadMlrLearner}
\alias{uploadMlrLearner}
\title{Upload an mlr learner to the server.}
\usage{
uploadMlrLearner(lrn, hash, dir = getwd(), delete.source = TRUE, ...,
  show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{lrn}{[\code{\link[mlr]{Learner}}]\cr
The mlr learner.}

\item{hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{dir}{[\code{character(1)}]\cr
A directory where to save the sourcefile. Default is the current working directory.}

\item{delete.source}{[\code{logical(1)}]\cr
Should the autogenerated source file be removed at the end? Default is \code{TRUE}.}

\item{...}{[\code{any}]\cr
Further arguments that are passed to \code{\link{createOpenMLImplementationForMlrLearner}}.
This could be an optional description of the learner or a name for the implementation. By
default, these are autogenerated.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\value{
[\code{integer(1)}]. Unique implementation ID.
}
\description{
This is a convenience function that generates an OpenML implementation (flow) as
  well as a sourcefile for your mlr learner. If the learner is not stored on the OpenML server
  yet, it is uploaded. The flow's implementation ID is returned in any case.
}
\seealso{
\code{\link[mlr]{makeLearner}}, \code{\link{authenticateUser}},
  \code{\link{uploadOpenMLImplementation}}
}

