% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uploadOMLDataSet.R
\name{uploadOMLDataSet}
\alias{uploadOMLDataSet}
\title{Upload an OpenML dataset to the server.}
\usage{
uploadOMLDataSet(data.set, data.file, verbosity = NULL)
}
\arguments{
\item{data.set}{[\code{\link{data.frame}}]\cr
The dataset that should be uploaded.}

\item{data.file}{[\code{character(1)}]\cr
Data file.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{invisible(numeric(1))}]. The id of the data (\code{did}).
}
\description{
Upload an OpenML dataset to the server.
}
\note{
Assumes that the given dataset is ARFF.
Maybe we can add dataframe to ARFF conversion here?
}

