% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uploadOMLFlow.R
\name{uploadOMLFlow}
\alias{uploadOMLFlow}
\title{Upload an OpenML flow to the server.}
\usage{
uploadOMLFlow(x, session.hash, verbosity, sourcefile, binaryfile)
}
\arguments{
\item{x}{[\code{\link{OMLFlow}}|\code{\link{Learner}}]\cr
The flow that should be uploaded.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}

\item{sourcefile}{[\code{character(1)}]\cr
the file path to the flow (not needed for \code{\link{Learner}})}

\item{binaryfile}{[\code{character(1)}]\cr
the file path to the flow (not needed for \code{\link{Learner}})}
}
\value{
[\code{invisible(numeric(1))}]. The id of the flow (\code{implementation.id}).
}
\description{
Upload an OpenML flow to the server.
}

