% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uploadOMLRun.R
\name{uploadOMLRun}
\alias{uploadOMLRun}
\title{Upload an OpenML run to the server.}
\usage{
uploadOMLRun(run, implementation.id, session.hash = getSessionHash(),
  verbosity = NULL)
}
\arguments{
\item{run}{[\code{\link{OMLRun}}]\cr
The run that should be uploaded.}

\item{implementation.id}{[\code{integer(1)}]\cr
ID of the implementation of an OpenML flow.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{invisible(numeric(1))}]. The run ID.
}
\description{
Share a run of an implementation on a given OpenML task.
}

