% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uploadOpenMLImplementation}
\alias{uploadOpenMLImplementation}
\title{Upload an OpenML implementation to the server.}
\usage{
uploadOpenMLImplementation(implementation, sourcefile, binaryfile, session.hash,
  delete.source.binary = FALSE, show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{implementation}{[\code{\link{OpenMLImplementation}}]\cr
An OpenMLImplementation object. Should at least contain a name and a description.}

\item{sourcefile}{[\code{character(1)}]\cr
The source code of the implementation. If multiple files, please zip them.
Either source or binary is required.}

\item{binaryfile}{[\code{character(1)}]\cr
The binary of the implementation. If multiple files, please zip them.
Either source or binary is required.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{delete.source.binary}{[\code{logical(1)}]\cr
Should all source and binary files be removed at the end? Default is \code{FALSE}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\description{
Upload an OpenML implementation to the server.
}

