% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/config.R
\name{configuration}
\alias{configuration}
\title{OpenML configuration.}
\description{
In order to understand how the package should be configured
please read
\url{https://github.com/tudo-r/BatchJobs/wiki/Configuration}.
}
\seealso{
Other config: \code{\link{authenticateUser}};
  \code{\link{getOMLConfig}}; \code{\link{getSessionHash}};
  \code{\link{setOMLConfig}}
}

