% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configuration}
\alias{configuration}
\title{OpenML configuration.}
\description{
After loading the package, it tries to find a configuration in your home
directory. The R command \code{path.expand("~/.openml/config")} gives you the
full path to the configuration file on your operating system.

For further information please read the configuration section in
\url{https://github.com/openml/r/wiki}.
}
\seealso{
Other config: \code{\link{getOMLConfig}},
  \code{\link{loadOMLConfig}}, \code{\link{saveOMLConfig}},
  \code{\link{setOMLConfig}}
}

