% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configuration}
\alias{configuration}
\title{OpenML configuration.}
\description{
After loading the package, it tries to find a configuration in your home
directory. The R command \code{path.expand("~/.openml/config")} gives you the
full path to the configuration file on your operating system.

For further information please read the
\href{https://github.com/openml/openml-r/blob/master/todo/OpenML.Rmd}{vignette}.
}
\note{
By default the cache directory is located in a temporary directory and
the cache will be deleted in between R sessions. We thus recommend to set
the cache directory by hand.
}
\seealso{
Other config: \code{\link{getOMLConfig}},
  \code{\link{loadOMLConfig}}, \code{\link{saveOMLConfig}},
  \code{\link{setOMLConfig}}
}
