% Generated by roxygen2 (4.0.2): do not edit by hand
\name{configureOpenML}
\alias{configureOpenML}
\title{Configures the behavior of the package.}
\usage{
configureOpenML(show.info)
}
\arguments{
\item{show.info}{[\code{logical(1)}]\cr
Some methods of OpenML support a \code{show.info} argument to enable
verbose output on the console. This option sets the default value for these arguments.
Setting the argument manually in one of these functions will overwrite the default
value for that specific function call.
Default is \code{TRUE}.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
Configuration is done by setting custom \code{\link{options}}.

If you do not set an option here, its current value will be kept.

If you call this function with an empty argument list, everything is set to its defaults.
}
\seealso{
Other configure: \code{\link{getOpenMLOptions}}
}

