% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMlrLearnerToOMLFlow.R
\name{convertMlrLearnerToOMLFlow}
\alias{convertMlrLearnerToOMLFlow}
\title{convertMlrLearnerToOMLFlow}
\usage{
convertMlrLearnerToOMLFlow(lrn, name = paste0("mlr.", lrn$id),
  description = NULL, ...)
}
\arguments{
\item{lrn}{[\code{\link[mlr]{Learner}}]\cr
The mlr learner.}

\item{name}{[\code{character(1)}]\cr
The name of the flow object. Default is the learner ID.}

\item{description}{[\code{character(1)}]\cr
An optional description of the learner.
Default is a short specification of the learner and the associated package.}

\item{...}{[\code{any}]\cr
Further optional parameters that are passed to \code{\link{makeOMLFlow}}.}
}
\value{
[\code{\link{OMLFlow}}].
}
\description{
Creates an OMLFlow for an mlr learner.
Required if you want to upload an mlr learner.
}

