% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOMLTaskToMlr.R
\name{convertOMLTaskToMlr}
\alias{convertOMLTaskToMlr}
\title{Convert an OpenML task to mlr.}
\usage{
convertOMLTaskToMlr(obj, ignore.flagged.attributes = TRUE,
  drop.levels = TRUE, verbosity = NULL)
}
\arguments{
\item{obj}{[\code{\link{OMLTask}}]\cr
The object that should be converted.}

\item{ignore.flagged.attributes}{[\code{logical(1)}]\cr
Should those features that are listed in the data set description slot \dQuote{ignore.attribute}
be removed?
Default is \code{TRUE}.}

\item{drop.levels}{[\code{logical(1)}]\cr
Drop empty factor levels in data?
Default is \code{TRUE}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
A list of:
  \item{mlr.task}{[\code{\link[mlr]{Task}}]}
  \item{mlr.rin}{[\code{\link[mlr]{ResampleInstance}}]}
  \item{mlr.measures}{[list of \code{\link[mlr]{Measure}s} to optimize for.}
}
\description{
Converts an \code{\link{OMLTask}} to a list of \code{\link[mlr]{Task}},
\code{\link[mlr]{ResampleInstance}} and \code{\link[mlr]{Measure}}.
}
\seealso{
Other task related functions: \code{\link{deleteOMLObject}},
  \code{\link{getOMLTask}}, \code{\link{listOMLTaskTypes}},
  \code{\link{listOMLTasks}}, \code{\link{makeOMLTask}}
}

