% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createOpenMLImplementationForMlrLearner}
\alias{createOpenMLImplementationForMlrLearner}
\title{createOpenMLImplementationForMlrLearner.}
\usage{
createOpenMLImplementationForMlrLearner(lrn, name = lrn$id, description, ...)
}
\arguments{
\item{lrn}{[\code{\link[mlr]{Learner}}]\cr
The mlr learner.}

\item{name}{[\code{character(1)}]\cr
The name of the implementation object. Default is the learner's ID.}

\item{description}{[\code{character(1)}]\cr
An optional description of the learner.
Default is a short specification of the learner and the associated package.}

\item{...}{[\code{any}]\cr
Further optional parameters that are passed to \code{\link{makeOpenMLImplementation}}.}
}
\value{
[\code{\link{OpenMLImplementation}}].
}
\description{
Create an OpenML implementation description object for an mlr learner.
Required if you want to upload an mlr learner.
}

