% Generated by roxygen2 (4.0.1): do not edit by hand
\name{downloadOpenMLDataAsMlrTask}
\alias{downloadOpenMLDataAsMlrTask}
\title{Download an OpenML data set and convert it into an mlr task.}
\usage{
downloadOpenMLDataAsMlrTask(name, dir = tempdir(), clean.up = TRUE,
  show.info = TRUE)
}
\arguments{
\item{name}{[\code{character}]\cr
The name of the data set.}

\item{dir}{[\code{character}]\cr
The directory where intermediate files are stored. If clean.up = TRUE, this does not matter.
Default is the session's temporary directory.}

\item{clean.up}{[\code{loigcal}]\cr
Should the downloaded files be removed from disk at the end?
Default is \code{TRUE}.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}
}
\value{
[\code{\link[mlr]{SupervisedTask}}]
}
\description{
Download an OpenML data set and convert it into an mlr task.
}

