% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadOpenMLRunResults}
\alias{downloadOpenMLRunResults}
\title{Download OpenML run results from the server.}
\usage{
downloadOpenMLRunResults(id, ignore.cache = FALSE,
  show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{id}{[\code{numeric}]\cr
The task ID.}

\item{ignore.cache}{[\code{logical(1)}]\cr
Ignore a possibly cached version and force
re-download of files? Default is \code{FALSE}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}

\item{clean.up}{[\code{logical(1)}]\cr
Should the downloaded run xml file be removed at the end?
Default is \code{TRUE}.}
}
\value{
[\code{\link{OpenMLRunResults}}]
}
\description{
Download OpenML run results from the server.
}

