% Generated by roxygen2 (4.0.1): do not edit by hand
\name{downloadOpenMLTask}
\alias{downloadOpenMLTask}
\title{Download a task from the OpenML repository.}
\usage{
downloadOpenMLTask(id, dir = tempdir(), clean.up = TRUE,
  fetch.data.set.description = TRUE, fetch.data.set = TRUE,
  fetch.data.splits = TRUE, show.info = TRUE)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
ID number of task on OpenML server, used to retrieve the task.}

\item{dir}{[\code{character(1)}]\cr
Directory where downloaded files from the repository are stored.
Default is the path of the per-session temporary directory \code{tempdir()}.}

\item{clean.up}{[\code{logical(1)}]\cr
Should the downloaded files be removed from disk at the end?
Default is \code{TRUE}.}

\item{fetch.data.set.description}{[\code{logical(1)}]\cr
Should the data set description also be downloaded?
Default is \code{TRUE}.}

\item{fetch.data.set}{[\code{logical(1)}]\cr
Should the data set also be downloaded?
Default is \code{TRUE}.}

\item{fetch.data.splits}{[\code{logical(1)}]\cr
Should the data splits (for resampling) also be downloaded?
Default is \code{TRUE}.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}
}
\value{
\code{\linkS4class{OpenMLTask}} object.
}
\description{
This function downloads an OpenML task and all associated files from the OpenML repository,
intermediately stores the files on disk and creates an S4 object which completely specifies the task.
}
\details{
Usually there is no reason to set the \code{fetch.*} arguments to \code{FALSE}, as you want all information
completely encapsulated in the task object.
}
\examples{
# Download task and access relevant information to start running experiments
\dontrun{
task <- downloadOpenMLTask(id = 1)
show(task)
print(task<at>task.type)
print(task<at>task.target.features)
print(head(task<at>data.desc<at>data.set))
}
}

