% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractOMLStudyIds.R
\name{extractOMLStudyIds}
\alias{extractOMLStudyIds}
\title{Extract IDs of a OMLStudy object}
\usage{
extractOMLStudyIds(object, type)
}
\arguments{
\item{object}{[\code{OMLStudy}] \cr
The OMLStudy object.}

\item{type}{[\code{character(1)}] \cr
A character that specifies which ids should be extracted from the study.
Can be either "data.id", "task.id", "flow.id" or "run.id".}
}
\value{
[\code{numeric}].
}
\description{
Extracts either all \code{data.id}s, \code{task.id}s, \code{flow.id}s or \code{run.id}s from an \code{OMLStudy} object.
}
