% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config_helpers.R, R/getOMLConfig.R
\name{getOMLConfig}
\alias{getOMLConfig}
\title{Returns a list of OpenML configuration settings.}
\usage{
getOMLConfig()

getOMLConfig()
}
\value{
\code{list} of current configuration variables with class \dQuote{OMLConfig}.

\code{list} of current configuration variables with class \dQuote{OMLConfig}.
}
\description{
Returns a list of OpenML configuration settings.

Returns a list of OpenML configuration settings
}
\seealso{
Other config: \code{\link{configuration}};
  \code{\link{loadOMLConfig}}; \code{\link{saveOMLConfig}};
  \code{\link{setOMLConfig}}

Other config: \code{\link{configuration}};
  \code{\link{loadOMLConfig}}; \code{\link{saveOMLConfig}};
  \code{\link{setOMLConfig}}
}

