% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOMLConfig.R
\name{getOMLConfig}
\alias{getOMLConfig}
\title{Get OpenML configuration.}
\usage{
getOMLConfig()
}
\value{
\code{list} of current configuration variables with class \dQuote{OMLConfig}.
}
\description{
Returns a list of OpenML configuration settings.
}
\examples{
getOMLConfig()
}
\seealso{
Other config: \code{\link{configuration}},
  \code{\link{loadOMLConfig}}, \code{\link{saveOMLConfig}},
  \code{\link{setOMLConfig}}
}
\concept{config}
