% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOMLDataSet.R
\name{getOMLDataSet}
\alias{getOMLDataSet}
\title{Get an OpenML data set.}
\usage{
getOMLDataSet(did, verbosity = NULL)
}
\arguments{
\item{did}{[\code{integer(1)}]\cr
Data set ID.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{\link{OMLDataSet}}].
}
\description{
Given a data set ID, the corresponding \code{\link{OMLDataSet}} will be downloaded (if not in cache)
and returned.

Note that data splits and other task-related information are not included in an \code{\link{OMLDataSet}}.
Tasks can be downloaded with \code{\link{getOMLTask}}.
}

