% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOMLDataSet}
\alias{getOMLDataSet}
\title{Get an OpenML data set.}
\usage{
getOMLDataSet(id, session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
The id of the data set.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{\link{OMLDataSet}}]
}
\description{
Given the id, this function will download the whole data set, including the ARFF and
the description XML from the OpenML server. Note that this does not include data
splits or other task-related information. Tasks can be downloaded with
\code{\link{downloadOMLTask}}.
}

