% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOMLFlow.R
\name{getOMLFlow}
\alias{getOMLFlow}
\title{Get an OpenML flow.}
\usage{
getOMLFlow(implementation.id, verbosity = NULL)
}
\arguments{
\item{implementation.id}{[\code{integer(1)}]\cr
ID of the implementation of an OpenML flow.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{\link{OMLFlow}}].
}
\description{
Given an implementation ID, the corresponding \code{\link{OMLFlow}} is downloaded
if not already available in cache.
}

