% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOMLFlow}
\alias{getOMLFlow}
\title{Download an OpenML flow.}
\usage{
getOMLFlow(id, session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
The ID of the desired flow.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{\link{OMLFlow}}].
}
\description{
Download an OpenML flow.
}

