% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOMLPredictions.R
\name{getOMLPredictions}
\alias{getOMLPredictions}
\title{Get predictions of an OpenML run.}
\usage{
getOMLPredictions(run, verbosity = NULL)
}
\arguments{
\item{run}{[\code{\link{OMLRun}}]\cr
The OpenML run.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}]
}
\description{
This function will return the predictions of a certain OpenML run.
}
\seealso{
\code{\link{OMLRun}}, \code{\link{getOMLRun}}
}

