% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getOMLRun.R
\name{getOMLRun}
\alias{getOMLRun}
\title{Get an OpenML run.}
\usage{
getOMLRun(id, session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
The task ID.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{\link{OMLRun}}]
}
\description{
This function downloads an OpenML run including all server and user computed metrics.
The associated predictions can be retrieved with \code{\link{getOMLPredictions}}.
}
\seealso{
To retrieve the corresponding predictions: \code{\link{getOMLPredictions}}
}

