% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOMLTask.R
\name{getOMLTask}
\alias{getOMLTask}
\title{Get an OpenML task.}
\usage{
getOMLTask(task.id, verbosity = NULL)
}
\arguments{
\item{task.id}{[\code{integer(1)}]\cr
The task ID.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}
}
\value{
[\code{\link{OMLTask}}]
}
\description{
This function downloads an OpenML task and all associated files from the OpenML repository,
caches the files on disk and creates an S3 object which completely specifies the task.
}
\examples{
# Download task and access relevant information to start running experiments
\dontrun{
task = getOMLTask(1)
print(task)
print(task$task.type)
print(task$input$data.set)
print(head(task$input$data.set$data))
}
}

