% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOpenMLDataQualityList}
\alias{getOpenMLDataQualityList}
\title{Get all names of OpenML data qualities.}
\usage{
getOpenMLDataQualityList(show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\description{
Get all names of OpenML data qualities.
}

