% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOpenMLDatasetNames}
\alias{getOpenMLDatasetNames}
\title{Get all names of OpenML data sets which are registered on the server.}
\usage{
getOpenMLDatasetNames()
}
\value{
[\code{data.frame}].
}
\description{
We currently return only data sets which are marked as safe on OpenML.
}

