% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOpenMLRegisteredTasks}
\alias{getOpenMLRegisteredTasks}
\title{Get information about all tasks registred on OpenML server.}
\usage{
getOpenMLRegisteredTasks(type = "classification")
}
\arguments{
\item{type}{[\code{character(1)}]\cr
  Currently only "classification" is supported.}
}
\value{
[\code{data.frame}].
}
\description{
We currently return only  which are marked as safe on OpenML.
}

