% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSessionHash}
\alias{getSessionHash}
\title{Get the current session's hash.}
\usage{
getSessionHash()
}
\value{
[\code{character(1)}].
}
\description{
Get the current session's hash.
}
\seealso{
Other config: \code{\link{authenticateUser}};
  \code{\link{configuration}}; \code{\link{getOMLConfig}};
  \code{\link{setOMLConfig}}
}

