% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTasksForDataSet}
\alias{getTasksForDataSet}
\title{Get task IDs for a certain data set}
\usage{
getTasksForDataSet(did)
}
\arguments{
\item{did}{[\code{integer}]\cr
One or more data set IDs.}
}
\value{
[\code{integer}]. A vector containing the task IDs.
}
\description{
Get task IDs for a certain data set
}

