% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listOMLDataSetQualities.R
\name{listOMLDataSetQualities}
\alias{listOMLDataSetQualities}
\title{List available OpenML qualities names.}
\usage{
listOMLDataSetQualities(verbosity = NULL)
}
\arguments{
\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned \code{data.frame} contains quality name \dQuote{name}.
}
\seealso{
Other list: \code{\link{getOMLDataSetQualities}};
  \code{\link{listOMLDataSets}};
  \code{\link{listOMLEvaluationMeasures}};
  \code{\link{listOMLFlows}}
}

