% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listOMLEstimProcs.R
\name{listOMLEstimProcs}
\alias{listOMLEstimProcs}
\title{List available Estimation Procedures}
\usage{
listOMLEstimProcs(session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned \code{data.frame} contains the \code{est.id} and the corresponding
name of the estimation procedure.
}

