% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listOMLRunEvaluations.R
\name{listOMLRunEvaluations}
\alias{listOMLRunEvaluations}
\title{List run results of a task.}
\usage{
listOMLRunEvaluations(task.id = NULL, setup.id = NULL, flow.id = NULL,
  run.id = NULL, uploader.id = NULL, verbosity = NULL)
}
\arguments{
\item{task.id}{[\code{integer(1)}]\cr
Task ID.}

\item{setup.id}{[\code{integer(1)}]\cr
ID of the parameter setup.}

\item{flow.id}{[\code{integer(1)}]\cr
ID of the implementation of an OpenML flow.}

\item{run.id}{[\code{integer}]\cr
a single ID or a vector of IDs of the runs.}

\item{uploader.id}{[\code{integer(1)}]\cr
ID of the uploader.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
Retrieves all run results for a task specified by \code{task.id} and returns
it in a \code{data.frame}. Each row contains, among others, the run id \dQuote{rid},
the setup id \dQuote{sid} and the task.id \dQuote{tid}.
}

