% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listOMLRuns}
\alias{listOMLRuns}
\title{List OpenML runs.}
\usage{
listOMLRuns(task.id = NULL, setup.id = NULL, impl.id = NULL,
  session.hash = getSessionHash(), verbosity = NULL)
}
\arguments{
\item{task.id}{[\code{integer}]\cr
Id of the task.}

\item{setup.id}{[\code{integer(1)}]\cr
Id of the parameter setup (?).}

\item{impl.id}{[\code{integer(1)}]\cr
Id of the implementation.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{data.frame}].
}
\description{
This function returns information on all OpenML runs that match a certain
\code{task.id}, \code{setup.id} and/or implementation ID \code{impl.id}.
}

