% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listOMLTasks.R
\name{listOMLTasks}
\alias{listOMLTasks}
\title{List available OpenML tasks}
\usage{
listOMLTasks(type = 1L, session.hash = getSessionHash(), verbosity = NULL,
  status = "active")
}
\arguments{
\item{type}{[\code{integer(1)}]\cr
The task type you want to list tasks for.
See code \link{listOMLTaskTypes}.
Default is \code{1} for classification.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}. Default is
\code{\link{getSessionHash}()}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}

\item{status}{[\code{character}]\cr
Subsets the results according to the status, can take one or more values of
\code{{"active", "deactivated", "in_preparation"}}.
Default is \code{"active"}.}
}
\value{
[\code{data.frame}].
}
\description{
The returned \code{data.frame} contains the \code{task_id}, the data set id \code{did},
the \code{status} and some describing data qualities.
}

