% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOMLConfig.R
\name{loadOMLConfig}
\alias{loadOMLConfig}
\title{load OpenML configuration}
\usage{
loadOMLConfig(path = "~/.openml/config", assign = TRUE)
}
\arguments{
\item{path}{[\code{character(1)}]\cr
Full path location of the config file to be loaded.}

\item{assign}{[\code{logical(1)}]\cr
Use the loaded configuration as the current configuration? Default is \code{TRUE}.
If set to \code{FALSE}, the configuration is just returned by the function.}
}
\value{
\code{list} of current configuration variables with class \dQuote{OMLConfig}.
}
\description{
Loads the openml config file from the disk and overwrites the current OpenML config.
}
\seealso{
Other config: \code{\link{configuration}},
  \code{\link{getOMLConfig}}, \code{\link{saveOMLConfig}},
  \code{\link{setOMLConfig}}
}

