% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLTask_Class.R
\name{makeOMLTask}
\alias{OMLTask}
\alias{makeOMLTask}
\title{Construct OMLTask.}
\usage{
makeOMLTask(task.id, task.type, input, parameters = list(), output,
  tags = NA_character_)
}
\arguments{
\item{task.id}{[\code{integer(1)}]\cr
The ID of this task. Generated by the API.}

\item{task.type}{[\code{character(1)}]\cr
The task type of this task. Task types can be browsed and created on the OpenML website.
See also \code{\link{listOMLTaskTypes}} for a list of all available tasks.}

\item{input}{[\code{character}]\cr
The inputs given for this task (data.set, estimation.procedure, evaluation.measures, cost.matrix).}

\item{parameters}{[\code{list}]\cr
Parameter settings for this task (depends on the task type).}

\item{output}{[\code{list}]\cr
Outputs expected after running this task.}

\item{tags}{[\code{character}]\cr
Optional tags describing the (data of the) task.}
}
\description{
More details about the elements of a \code{OMLTask} can be found in the
\href{https://github.com/openml/website/blob/master/openml_OS/views/pages/api_new/v1/xsd/openml.task.search.xsd}{XSD scheme}.
}
\seealso{
Other task related functions: \code{\link{convertOMLTaskToMlr}},
  \code{\link{deleteOMLObject}}, \code{\link{getOMLTask}},
  \code{\link{listOMLTaskTypes}},
  \code{\link{listOMLTasks}}
}

