% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeOpenMLTask}
\alias{OpenMLTask}
\alias{makeOpenMLTask}
\title{Construct OpenMLTask.}
\usage{
makeOpenMLTask(task.id, task.type, task.pars = list(),
  task.target.features = NA_character_, task.data.desc.id,
  task.data.desc = NULL, task.estimation.procedure, task.preds = list(),
  task.evaluation.measures)
}
\arguments{
\item{task.id}{[\code{integer(1)}]\cr
The ID of this task. Generated by the API.}

\item{task.type}{[\code{character(1)}]\cr
The task type of this task. Task types can be browsed and created on the OpenML website.}

\item{task.pars}{[\code{list}]\cr
A set of parameters specific to the task type.}

\item{task.target.features}{[\code{character}]\cr
The name(s) of the target feature(s). There may be none, one or multiple target features.}

\item{task.data.desc.id}{[\code{integer(1)}]\cr
The OpenML ID of the data set associated with the task.}

\item{task.data.desc}{[\code{\link{OpenMLDataSetDescription}}]\cr
Information on the data set. Optional.}

\item{task.estimation.procedure}{[\code{\link{OpenMLEstimationProcedure}}]\cr
Information on the task's estimation method and the asoociated data splits. Optional.}

\item{task.preds}{[\code{list}]\cr
A list that contains information on the format of the predictions for the particular task.}

\item{task.evaluation.measures}{[\code{character}]\cr
The evaluation measure(s) that should be used for optimization.}
}
\description{
Construct OpenMLTask.
}
\seealso{
\code{\linkS4class{OpenMLDataSetDescription}}
}

