% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeOpenMLTaskResults}
\alias{OpenMLTaskResults}
\alias{makeOpenMLTaskResults}
\title{Construct OpenMLTaskResults.}
\usage{
makeOpenMLTaskResults(task.id, task.name, task.type.id = NA_integer_,
  input.data = NA_integer_, estimation.procedure = NA_character_, metrics)
}
\arguments{
\item{task.id}{[\code{integer(1)}]\cr
ID of the OpenMLTask.}

\item{task.name}{[\code{character(1)}]\cr
A name describing the task shortly.}

\item{task.type.id}{[\code{integer(1)}]\cr
ID of the task type.}

\item{input.data}{[\code{integer(1)}]\cr
ID of the data set that belongs to the task.}

\item{estimation.procedure}{[\code{character(1)}]\cr
The task's estimation procedure.}

\item{metrics}{[\code{data.frame}]\cr
A data.frame of the metrics of all runs that were uploaded for this task.}
}
\description{
Objects of this class are returned by \code{\link{downloadOpenMLTaskResults}}. All
  members are filled in by the server.
}
\details{
OpenMLTaskResults
}
\seealso{
\code{\link{downloadOpenMLTaskResults}}
}

