% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getOMLDataSet.R
\name{OMLDataSet}
\alias{OMLDataSet}
\title{OMLDataSet.}
\description{
An \code{OMLDataSet} consists of an \code{OMLDataSetDescription}, a
\code{data.frame} containing the data set and, finally, old and new column names.

The \code{OMLDataSetDescription} provides information on the data set, like the ID, name,
version, etc. To see a full list of all elements, please see the
\href{https://github.com/openml/website/blob/master/openml_OS/views/pages/rest_api/xsd/openml.data.upload.xsd}{XSD}.

The slot \code{colnames.old} contains the original names, i.e., the column names that were
uploaded to the server, while \code{colnames.new} contains the names that you will see when
working with the data in R.
Most of the time, old and new column names are identical. Only if the original names are
not valid, the new ones will differ.
}

