% Generated by roxygen2 (4.0.1): do not edit by hand
\name{OptionalOpenMLEstimationProcedure}
\alias{OptionalOpenMLEstimationProcedure}
\alias{OptionalOpenMLEstimationProcedure-class}
\title{OptionalOpenMLEstimationProcedure}
\description{
Either an object of class \code{\link{OpenMLEstimationProcedure}} or \code{NULL}.
}
\seealso{
\code{\link{OpenMLEstimationProcedure}}, \code{\link{OpenMLTask}}
}

