\name{OptionalOpenMLEstimationProcedure}
\alias{OptionalOpenMLEstimationProcedure}
\alias{OptionalOpenMLEstimationProcedure-class}
\title{OptionalOpenMLEstimationProcedure}
\description{
Either an object of class
\code{\link{OpenMLEstimationProcedure}} or \code{NULL}.
}
\seealso{
\code{\link{OpenMLEstimationProcedure}},
\code{\link{OpenMLTask}}
}

