% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/populateCache.R
\name{populateOMLCache}
\alias{populateOMLCache}
\title{Download a bunch of OpenML objects to cache.}
\usage{
populateOMLCache(task.ids = NULL, flow.ids = NULL, run.ids = NULL,
  dids = NULL, verbosity = NULL)
}
\arguments{
\item{task.ids}{[\code{integer}]\cr
Vector of task ids.}

\item{flow.ids}{[\code{integer}]\cr
Vector of flow ids.}

\item{run.ids}{[\code{integer}]\cr
Vector of run ids.}

\item{dids}{[\code{integer}]\cr
Vector of dataset ids.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{invisible(NULL)}]
}
\description{
Given a set of tasks ids, flow ids, run ids and/or dataset ids
the function populates the cache directory by downloading the corresponding
objects. This is of particular interest in highly parallel computations on
a cluster with a shared file system.
}

