% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runSQLQuery}
\alias{runSQLQuery}
\title{Run an SQL query on the OpenML DB and retrieve results.}
\usage{
runSQLQuery(query, simplify = TRUE,
  show.info = getOpenMLOption("show.info"))
}
\arguments{
\item{query}{[\code{character(1)}]\cr
The SQL query.}

\item{simplify}{[\code{logical(1)}]\cr
If there is only one column in the resulting table,
directly return the corresponding vector?
Default is \code{TRUE}.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureOpenML}}.}
}
\value{
[\code{data.frame}]. The results as a table.
}
\description{
Most users should not need to use this, only a last
measure to access all information on the server.
}

