% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runTask}
\alias{runTask}
\title{Solve a task with mlr.}
\usage{
runTask(task, learner, return.mlr.results = FALSE,
  remove.const.feats = TRUE, ..., verbosity = NULL)
}
\arguments{
\item{task}{[\code{\link{OMLTask}}]\cr
An OpenML task. Required.}

\item{learner}{[\code{\link[mlr]{Learner}}]\cr
Learner object from package mlr to solve the task.}

\item{return.mlr.results}{[\code{logical(1)}]\cr
Should not only the predictions but all of the by mlr computed information be returned?
This includes test measures in each step of the resampling procedure as well as the aggregated
performance. See \code{\link[mlr]{resample}}. Default is \code{TRUE}.}

\item{remove.const.feats}{[\code{logical(1)}]\cr
Should constant features be removed? Default is \code{TRUE}. Note, that setting this
to \code{FALSE} may lead to fatal errors.}

\item{...}{[any]\cr
Further arguments that are passed to \code{\link[mlr]{removeConstantFeatures}}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
List of:
  \item{run.pred}{[\code{\link[mlr]{ResamplePrediction}}]\cr
    Predictions resulting from the run. These are necessary in order to upload a run.}
  \item{mlr.resample.results}{[\code{list}]\cr
    The results of the mlr function \code{\link[mlr]{resample}}.\cr
    \code{NULL} if \code{return.mlr.results == FALSE}.}
}
\description{
Solve a task with a specified learner from mlr and receive predictions that are capable
of being uploaded to the OpenML server.
}
\seealso{
\code{\link{OMLTask}}, \code{\link[mlr]{learners}},
  \code{\link{authenticateUser}}, \code{\link[mlr]{resample}}
}

