% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runTaskMlr.R
\name{runTaskMlr}
\alias{runTaskMlr}
\title{Run mlr learner on OpenML task.

Run task with a specified learner from mlr and produce predictions.}
\usage{
runTaskMlr(task, learner, remove.const.feats = TRUE, ..., verbosity = NULL)
}
\arguments{
\item{task}{[\code{\link{OMLTask}}]\cr
An OpenML task.}

\item{learner}{[\code{\link[mlr]{Learner}}]\cr
Learner from package mlr to run the task.}

\item{remove.const.feats}{[\code{logical(1)}]\cr
Should constant features be removed?
Default is \code{TRUE}.}

\item{...}{[any]\cr
Further arguments that are passed to \code{\link[mlr]{removeConstantFeatures}}.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{OMLMlrRun}], an \code{\link{OMLRun}} with an additional slot \code{mlr.resample.result}.
}
\description{
Run mlr learner on OpenML task.

Run task with a specified learner from mlr and produce predictions.
}
\seealso{
\code{\link{getOMLTask}}, \code{\link[mlr]{makeLearner}}
}

