\name{toMLR}
\alias{toMLR}
\title{Convert an OpenML task to an MLR task object.}
\usage{
toMLR(task)
}
\arguments{
  \item{task}{[\code{\linkS4class{OpenMLTask}}]\cr An
  OpenML task object. Required.}
}
\value{
[\code{\link[mlr]{SupervisedTask}}]
}
\description{
Convert an OpenML task to an MLR task object.
}

