% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toMlr}
\alias{toMlr}
\alias{toMlr.OpenMLDataSetDescription}
\alias{toMlr.OpenMLTask}
\title{Convert an OpenML object to mlr.}
\usage{
toMlr(obj, target, remove.target.NAs, ignore.flagged.attributes)

\method{toMlr}{OpenMLTask}(obj,
  target = obj$data.desc$default.target.attribute, remove.target.NAs = TRUE,
  ignore.flagged.attributes = TRUE)

\method{toMlr}{OpenMLDataSetDescription}(obj,
  target = obj$default.target.attribute, remove.target.NAs = TRUE,
  ignore.flagged.attributes = TRUE)
}
\arguments{
\item{obj}{[\code{\link{OpenMLDataSetDescription}} | \code{\link{OpenMLTask}}]\cr
The object that should be converted. Required.}

\item{target}{[\code{character}]\cr
The target for the classification/regression task. Default is the
\code{default.target.attribute} of the \code{DataSetDescription}.}

\item{remove.target.NAs}{[\code{logical(1)}]\cr
Should rows with missing target values be removed? Default is \code{TRUE}. Note, that
the function might fail if you set this to \code{FALSE}.}

\item{ignore.flagged.attributes}{[\code{logical(1)}]\cr
Should those features that are listed in the data set description's member "ignore.attribute"
be ignored? Default is \code{TRUE}.}
}
\value{
Either a [\code{\link[mlr]{Task}}] or, a list of:
  \item{mlr.task}{[\code{\link[mlr]{Task}}]\cr
    The task.}
  \item{mlr.rin}{[\code{\link[mlr]{ResampleInstance}}]\cr
    A server defined resample instance.}
  \item{mlr.measures}{[\code{list}]\cr
    A list of \code{\link[mlr]{Measure}s} to optimize for.}
  \item{original.lvls}{[\code{character}]\cr
    All original (possibly empty) levels of the target feature. Empty levels are
    removed from the data set during conversion but their names are needed to produce proper
    uploadable predictions.}
}
\description{
This function converts an \code{\link{OpenMLDataSetDescription}}
  or an \code{\link{OpenMLTask}} into an mlr \code{\link[mlr]{Task}} and, in case of a given
  \code{\link{OpenMLTask}}, various other mlr objects (see below).
}

