% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toMlr}
\alias{toMlr}
\title{Convert an OpenML task to an mlr task object.}
\usage{
toMlr(task)
}
\arguments{
\item{task}{[\code{\linkS4class{OpenMLTask}}]\cr
An OpenML task object. Required.}
}
\value{
[\code{\link[mlr]{Task}}]
}
\description{
Convert an OpenML task to an mlr task object.
}

