% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadOMLDataSet.R
\name{uploadOMLDataSet}
\alias{uploadOMLDataSet}
\title{Upload a dataset to the OpenML server.}
\usage{
uploadOMLDataSet(x, verbosity = NULL)
}
\arguments{
\item{x}{[\code{\link[mlr]{Task}}|[\code{\link{OMLDataSet}}]\cr
Contains the dataset that should be uploaded.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{invisible(numeric(1))}].
  The id of the data (\code{did}).
}
\description{
Share a dataset by uploading it to the OpenML server.
}
\seealso{
Other dataset related functions: \code{\link{OMLDataSetDescription}},
  \code{\link{OMLDataSet}},
  \code{\link{convertOMLDataSetToMlr}},
  \code{\link{deleteOMLObject}},
  \code{\link{getOMLDataSet}},
  \code{\link{listOMLDataSets}}

Other uploading functions: \code{\link{uploadOMLFlow}},
  \code{\link{uploadOMLRun}}
}

