% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadOMLRun.R
\name{uploadOMLRun}
\alias{uploadOMLRun}
\title{Upload an OpenML run.}
\usage{
uploadOMLRun(run, tags = NULL, verbosity = NULL, ...)
}
\arguments{
\item{run}{[\code{\link{OMLRun}}|\code{\link{runTaskMlr}}]\cr
The run that should be uploaded. Either a \code{\link{OMLRun}} or a run created with \code{\link{runTaskMlr}}.}

\item{tags}{[\code{character}]\cr
The tags that should be added after uploading.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}

\item{...}{Not used.}
}
\value{
[\code{invisible(numeric(1))}].
  The run ID.
}
\description{
Share a run of a flow on a given OpenML task by uploading it to the OpenML server.
}
\note{
This function will reset the cache of \code{\link{listOMLRuns}} and
\code{\link{listOMLRunEvaluations}} on success.
}
\seealso{
Other run-related functions: \code{\link{convertOMLRunToBMR}},
  \code{\link{deleteOMLObject}}, \code{\link{getOMLRun}},
  \code{\link{listOMLRuns}},
  \code{\link{makeOMLRunParameter}},
  \code{\link{makeOMLRun}}, \code{\link{tagOMLObject}}

Other uploading functions: \code{\link{uploadOMLDataSet}},
  \code{\link{uploadOMLFlow}}
}

