% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uploadOMLRun.R
\name{uploadOMLRun}
\alias{uploadOMLRun}
\title{Upload an OpenML run to the server.}
\usage{
uploadOMLRun(run, flow.id, verbosity = NULL)
}
\arguments{
\item{run}{[\code{\link{OMLRun}}]\cr
The run that should be uploaded.}

\item{flow.id}{[\code{integer(1)}]\cr
ID of the implementation of an OpenML flow.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{invisible(numeric(1))}] \cr
  The run ID.
}
\description{
Share a run of an flow on a given OpenML task.
}
\seealso{
Other run related functions: \code{\link{OMLRunParameter}},
  \code{\link{makeOMLRunParameter}}; \code{\link{OMLRun}},
  \code{\link{makeOMLRun}}; \code{\link{deleteOMLObject}};
  \code{\link{getOMLPredictions}}; \code{\link{getOMLRun}};
  \code{\link{listOMLRuns}}

Other uploading functions: \code{\link{uploadOMLDataSet}};
  \code{\link{uploadOMLFlow}}
}

