% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadOMLRun.R
\name{uploadOMLRun}
\alias{uploadOMLRun}
\title{Upload an OpenML run to the server.}
\usage{
uploadOMLRun(run, verbosity = NULL)
}
\arguments{
\item{run}{[\code{\link{OMLRun}}]\cr
The run that should be uploaded.}

\item{verbosity}{[\code{integer(1)}]\cr
Print verbose output on console? Possible values are:\cr
\code{0}: normal output,\cr
\code{1}: info output,\cr
\code{2}: debug output.\cr
Default is set via \code{\link{setOMLConfig}}.}
}
\value{
[\code{invisible(numeric(1))}]. 
  The run ID.
}
\description{
Share a run of an flow on a given OpenML task.
}
\seealso{
Other run related functions: \code{\link{deleteOMLObject}},
  \code{\link{getOMLRun}}, \code{\link{listOMLRuns}},
  \code{\link{makeOMLRun}}

Other uploading functions: \code{\link{uploadOMLDataSet}},
  \code{\link{uploadOMLFlow}}
}

