% Generated by roxygen2 (4.0.1): do not edit by hand
\name{uploadOpenMLRun}
\alias{uploadOpenMLRun}
\title{Upload an OpenML run to the server.}
\usage{
uploadOpenMLRun(task, mlr.lrn, impl.id, predictions, error.msg, session.hash,
  run.pars = NULL, show.info = TRUE)
}
\arguments{
\item{task}{[\code{\link{OpenMLTask}}]\cr
The task.}

\item{mlr.lrn}{[\code{\link[mlr]{Learner}}]\cr
The mlr learner, if used. Optional.}

\item{impl.id}{[\code{\link{character}}]\cr
The ID of the OpenML implementation that is stored on the OpenML server.}

\item{predictions}{[\code{data.frame}]\cr
The predictions. Must have the same form as produced by \code{\link{runTask}}. If a fatal error occured while
computing predictions, you have the possibility to leave this out and pass an error message instead (as parameter
error.msg).}

\item{error.msg}{[\code{try-error}]\cr
An optional error message. If this is not empty, the predictions (if present) are ignored.}

\item{session.hash}{[\code{character(1)}]\cr
A session token returned by \code{\link{authenticateUser}}.}

\item{run.pars}{[\code{list}]\cr
If no mlr learner was used, this is a list of all \code{\link{OpenMLRunParameter}s}.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}
}
\value{
[\code{numeric(1)} or \code{NULL}]. Run ID if the run was uploaded succesfully.
}
\description{
Share a run of an implementation on a given OpenML task.
}

